pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-rte"
version = "0.1.0"
description = "Python http client for RTE API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
  {name = "Enzo Boulin", email = "enzoboulin@gmail.com"}
]
dependencies = [
    "httpx>=0.28.1",
    "inline-snapshot>=0.31.1",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pytest>=9.0.0",
    "vcrpy>=7.0.0",
    "pydantic>=2.12.4",
]


[project.urls]
"Homepage" = "https://github.com/enzo-boulin/pyrte"
